Place the saved search checkboxes inside a ScrolledWindow
The ScrolledWindow will expand up to 300px if there are enough saved searches. However, the minimum size of the ScrolledWindow is 40px, even if the min_content_size is smaller.

Also updated pack_start's expand and fill params to ensure that the UI elements don't expand when the window size changes.